+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 2013.03.10, Version 0.1.2
 
  * Removes ProgressBar and change exporting indication
+ * Updates pass asking with getpass()